+++ epserde/tests/fail/drop_encase.rs
@@ -5,13 +5,12 @@
  */
 
 use epserde::{deser::Owned, prelude::*};
-use std::ops::Deref;
 
 fn main() {
     let v = vec![0, 10, 20, 30, 40];
+
     let mem_case = <MemCase<Owned<Vec<i32>>>>::encase(v);
-    let d = mem_case.deref();
-    let d = d.clone();
+    let u = mem_case.uncase();
     drop(mem_case);
-    let _d = d;
+    let _u = u;
 }